class5/demo-mcp-server/pyproject.toml
[project]
name = "demo-mcp-server"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
]
